Migrate Async storage mechanics to backend (#39)
* RD-8834 refactor: Expose API client to other service functions

* RD-8834 chore: Create type defs for My List and My List API responses

* RD-8834 feat: Use BE endpoints for my list operations

We can let the BE handle the business logic.

* RD-8834 refactor: Make the `useMyList` hook compatible with BE services

* RD-8834 refactor: Make consumers of My List operations compatible w/ BE

* RD-8834 refactor: Make `customerId` a number, instead of a string

* RD-8834 chore: Create type defs for Watch Status + API responses

* RD-8834 feat: Use BE for watch status operations

* RD-8834 refactor: Make consumers of Watch Status compatible w/ BE

* RD-8834 chore: Remove unnecessary cleanup calls

There were a couple of removal calls made to remove `my list` and `continue watching` items from
AsyncStorage, however, those are now obsolete since the persistence layer in the backend is
idempotent - no unnecessary space will be used.

* RD-8834 chore: Formatting + Remove unused imports

* RD-8834 chore: Create type defs for Search history + API responses

* RD-8834 feat: Use BE for search history operations

* RD-8834 refactor: Make consumers of Search History compatible w/ BE

* RD-8834 fix: Retrieve event IDs based on the found video IDs

The solution for retrieving all unique IDs in the `Continue Watching` section is to retrieve all
events whose video ID coincides with the ones found in `Continue Watching`.

* RD-8834 feat: Added video details retriever func to unique event getter

* RD-8834 chore: Formatting

* RD-8834 fix: Use video IDs from Diese, instead of those from Prismic

The IDs from Prismic change from video to video, whereas, those from Diese remain the same per
event.

* RD-8834 fix: Implemented `useContinueWatching` via a Diese-Event map

This is really not elegant, but this is the best solution I could find. Since there are many things
tightly coupled together, it wouldn't have been feasible to change the shape of the data in the
store. This solution is a bit of a hack, but works well enough!

* RD-8834 chore: Remove unnecessary console.log

* improved Andrey PR. Fixed some bugs

Co-authored-by: anton <19mann84@gmail.com>